Move OpaqueKeys to primitives.
