+++ core/sr-primitives/src/traits.rs
@@ -878,3 +878,75 @@ pub trait ValidateUnsigned {
 	/// Changes made to storage should be discarded by caller.
 	fn validate_unsigned(call: &Self::Call) -> TransactionValidity;
 }
+
+/// Opaque datatype that may be destructured into a series of raw byte slices (which represent
+/// individual keys).
+pub trait OpaqueKeys {
+	/// Return the number of encoded keys.
+	fn count() -> usize { 0 }
+	/// Get the raw bytes of key with index `i`.
+	fn get_raw(&self, i: usize) -> &[u8];
+	/// Get the decoded key with index `i`.
+	fn get<T: Decode>(&self, i: usize) -> Option<T> { T::decode(&mut self.get_raw(i)) }
+	/// Verify a proof of ownership for the keys.
+	fn ownership_proof_is_valid(&self, _proof: &[u8]) -> bool { true }
+}
+
+/// Calls a given macro a number of times with a set of fixed params and an incrementing numeral.
+/// e.g.
+/// ```
+/// count!(println ("{}",) foo, bar, baz);
+/// // Will result in three `println!`s: "0", "1" and "2".
+/// ```
+#[macro_export]
+macro_rules! count {
+	($f:ident ($($x:tt)*) ) => ();
+	($f:ident ($($x:tt)*) $x1:tt) => { $f!($($x)* 0); };
+	($f:ident ($($x:tt)*) $x1:tt, $x2:tt) => { $f!($($x)* 0); $f!($($x)* 1); };
+	($f:ident ($($x:tt)*) $x1:tt, $x2:tt, $x3:tt) => { $f!($($x)* 0); $f!($($x)* 1); $f!($($x)* 2); };
+	($f:ident ($($x:tt)*) $x1:tt, $x2:tt, $x3:tt, $x4:tt) => {
+		$f!($($x)* 0); $f!($($x)* 1); $f!($($x)* 2); $f!($($x)* 3);
+	};
+	($f:ident ($($x:tt)*) $x1:tt, $x2:tt, $x3:tt, $x4:tt, $x5:tt) => {
+		$f!($($x)* 0); $f!($($x)* 1); $f!($($x)* 2); $f!($($x)* 3); $f!($($x)* 4);
+	};
+}
+
+#[macro_export]
+/// Just implement `OpaqueKeys` for a given tuple-struct.
+/// Would be much nicer for this to be converted to `derive` code.
+macro_rules! impl_opaque_keys {
+	(
+		pub struct $name:ident ( $( $t:ty ),* $(,)* );
+	) => {
+		impl_opaque_keys! {
+			pub struct $name ( $( $t ,)* );
+			impl OpaqueKeys for _ {}
+		}
+	};
+	(
+		pub struct $name:ident ( $( $t:ty ),* $(,)* );
+		impl OpaqueKeys for _ {
+			$($rest:tt)*
+		}
+	) => {
+		#[derive(Default, Clone, PartialEq, Eq, Encode, Decode)]
+		#[cfg_attr(feature = "std", derive(Debug, Serialize, Deserialize))]
+		pub struct $name($( pub $t ,)*);
+		impl $crate::OpaqueKeys for $name {
+			fn count() -> usize {
+				let mut c = 0;
+				$( let _: $t; c += 1; )*
+				c
+			}
+			fn get_raw(&self, i: usize) -> &[u8] {
+				$crate::count!(impl_opaque_keys (!! self i) $($t),*);
+				&[]
+			}
+			$($rest)*
+		}
+	};
+	( !! $self:ident $param_i:ident $i:tt) => {
+		if $param_i == $i { return $self.$i.as_ref() }
+	}
+}
